+++ tofuref/data/helpers.py
@@ -1,11 +1,8 @@
 import json as jsonlib
 import logging
-import re
 from typing import Any
 
 import httpx
-from yaml import safe_load
-from yaml.scanner import ScannerError
 
 from tofuref import __version__
 from tofuref.config import config
@@ -14,25 +11,6 @@ from tofuref.data.cache import cached_file_path, get_from_cache, save_to_cache
 LOGGER = logging.getLogger(__name__)
 
 CODEBLOCK_REGEX = r"^```([a-z]+)\n([\s\S]*?)^```"
-
-
-def header_markdown_split(contents: str) -> tuple[dict, str]:
-    """
-    Most of the documentation files from the registry have a YAML "header"
-    that we mostly (at the moment) don't care about. Either way we
-    check for the header, and if it's there, we split it.
-    """
-    header = {}
-    if re.match(r"^---$", contents, re.MULTILINE):
-        split_contents = re.split(r"^---$", contents, maxsplit=2, flags=re.MULTILINE)
-        try:
-            header = safe_load(split_contents[1])
-        except ScannerError as _:
-            header = {}
-        markdown_content = split_contents[2]
-    else:
-        markdown_content = contents
-    return header, markdown_content
 
 
 async def get_registry_api(endpoint: str, json: bool = True, log_widget: Any | None = None) -> dict[str, dict] | str: